Move TestCancel into retry_test.go.
